pyproject.toml
[project]
name = "excel-interview-agent"
version = "0.1.0"
description = "Multi-agent Excel interview system using LangGraph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gradio>=5.46.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-google-genai>=2.1.12",
    "langgraph>=0.6.7",
    "python-dotenv>=1.1.1",
]
